Make `Receiver::sender` take `&mut self`
This fixes a soundness issue: because the implementation of
`Receiver::sender` does not use a Read-Modify-Write but a simple
Relaxed load to check if there is no live sender, it was theoretically
possible in some edges cases to create multiple senders from different
threads.

This commit also includes some minor refactoring in preparation for
upcoming features.